requirements.txt
agate==1.9.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
daff==1.4.2
dataclasses-json==0.6.7
dbt-adapters==1.15.3
dbt-common==1.25.0
dbt-core==1.9.6
dbt-duckdb==1.9.3
dbt-extractor==0.6.0
dbt-protos==1.0.317
dbt-semantic-interfaces==0.7.4
debugpy==1.8.14
decorator==5.2.1
deepdiff==7.0.1
distro==1.9.0
dnspython==2.7.0
duckdb==1.3.0
duckdb_engine==0.17.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.6.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isntall==1.0.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-experimental==0.3.4
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.44
leather==0.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mashumaro==3.14
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
openai==1.83.0
ordered-set==4.1.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pinecone==7.0.2
pinecone-client==6.0.0
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.41
sqlglot==26.25.3
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.9.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.3
google-auth==2.30.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
